Qarnot_ED_1/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Augment.py: 68
Augmentor == 0.2.8

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Augment.py: 74
# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Treatment.py: 1
# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 74
Pillow == 8.1.0

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 242
altair == 4.1.0

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Augment.py: 73
# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 73
matplotlib == 3.4.1

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Augment.py: 70
# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 70
numpy == 1.19.2

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 243
pandas == 1.2.2

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Launch_ED_on_Qarnot.py: 5
# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Launch_ED_part3_on_Qarnot.py: 5
qarnot == 2.2.4

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Augment.py: 71,72
# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 71,72
scikit_learn == 0.24.1

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 183,184,185,186
# tensorflow_gpu == 2.3.1

# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\Augment.py: 75
# D:\00_TRAV\PATATE\Qarnot_Wrapper-master\Qarnot_ED\encodDecod.py: 75
tqdm == 4.56.0
